requirements.txt
colorama
yara-python
requests